Added a basic ReportFormatter interface. Extracted the text report rendering code in APICoverageReport to a separate CoverageReportTextFormatter class, other minor refactoring.
