+++ junit5-support/src/main/kotlin/in/specmatic/test/formatters/ReportFormatter.kt
@@ -1,0 +1,5 @@
+package `in`.specmatic.test.formatters
+
+interface ReportFormatter<T> {
+    fun format(report: T): String
+}
